auto/lib/libxslt/conf: allow to override ngx_feature_path and ngx_feature_libs
Because libxml2 headers are not in /usr/include by default, hardcoding the
include directory to /usr/include/libxml2 does not play well when
cross-compiling, or if libxml2 has been installed somewhere else.

This patch allows to define/override the libxslt include directory, and
the libxslt libs flags.

Being able to override the include location is especially useful when
cross-compiling to prevent gcc from complaining about unsafe include
location for cross-compilation (-Wpoision-system-directories).

So far, this warning is only triggered by libxslt.

Signed-off-by: Samuel Martin <s.martin49@gmail.com>